Added Config

+++ db.py
@@ -38,11 +38,26 @@ class DB:
 
         self.create_global_table()
 
+        # Create Config Table as well
+        self.create_config_table()
+
         print("Database Created!")
 
     def create_global_table(self) -> None:
         self.cursor.execute(
             f"CREATE TABLE global(leaderboard VARCHAR(100) NOT NULL, symbol VARCHAR(50) DEFAULT ':coin:')")
+        self.db.commit()
+
+    def create_config_table(self) -> None:
+        self.cursor.execute(
+            "CREATE TABLE config(property VARCHAR(50), value VARCHAR(70))")
+        self.cursor.executemany(
+            "INSERT INTO config VALUES(%s, %s)", [
+                ["user_role", "bot-user"],
+                ["color_success", "0x0084ff"],
+                ["color_error", "0x991b1b"]
+            ]
+        )
         self.db.commit()
 
     def create_table(self, name: str, symbol: str = ":coin:") -> dict:
@@ -151,6 +166,9 @@ class DB:
 
         for table in tables:
             table_name = table[0]
+            if table_name == "config":
+                continue
+
             self.cursor.execute(f"DROP TABLE {table_name}")
         self.db.commit()
 
@@ -161,6 +179,52 @@ class DB:
             "message": "Deleted all the leaderboards"
         }
 
+    def get_config(self, option: str = "") -> dict:
+        query = "SELECT * FROM config"
+
+        if option != "":
+            query += f" WHERE property LIKE '{option}'"
+
+        self.cursor.execute(query)
+        values = self.cursor.fetchall()
+
+        if len(values) == 0:
+            return {
+                "success": False,
+                "message": f"Property `{option}` does not exist"
+            }
+
+        return {
+            "success": True,
+            "message": values
+        }
+
+    def set_config(self, option: str = "", value: str = "") -> dict:
+        if option == "" or value == "":
+            return {
+                "success": False,
+                "message": "Cannot set empty values"
+            }
+
+        self.cursor.execute(
+            f"SELECT * FROM config WHERE property LIKE '{option}'")
+        res = self.cursor.fetchall()
+
+        if len(res) == 0:
+            return {
+                "success": False,
+                "message": f"Property `{option}` does not exist"
+            }
+
+        self.cursor.execute(
+            f"UPDATE config SET value = '{value}' WHERE property LIKE '{option}'")
+        self.db.commit()
+
+        return {
+            "success": True,
+            "message": f"Set Property `{option}` to `{value}`"
+        }
+
     # Utils
 
     def get_symbol(self, leaderboard: str) -> str:
@@ -168,6 +232,13 @@ class DB:
             f"SELECT symbol from global WHERE leaderboard LIKE '{leaderboard}'")
         symbol = self.cursor.fetchone()[0]
         return symbol
+
+    def get_single_config(self, option: str = "") -> str:
+        self.cursor.execute(
+            f"SELECT value FROM config WHERE property LIKE '{option}'")
+        value = self.cursor.fetchone()[0]
+
+        return value
 
     def table_exists(self, table_name: str) -> bool:
         self.cursor.execute(f"SHOW TABLES LIKE '{table_name}'")
